Consolidating authenticator helpers
Moving Authenticator2 functionality into OAuth2Helper. Deleting
Authenticator2
